Add gulp task for generating NW.js apps
Also, now that we have support for regenerating all the files
in ./build/ properly, we can safely delete it from the repository
to clean up commits and PRs. They no longer need to update the
generated files in ./build/.

Generated apps are stored in ./apps/. Run gulp.js release to
generate apps for win32, osx64 and linux64.

To package the Chrome app, run gulp.js dist and pack the ./dist/
folder.