Organize OSD elements into groups
Implement groups which can be declared in any order including
any item in the order deemed more useful. This way we no longer
need to put elements ordered by their osd_items_e value.

Also, elements and even groups can be disabled when the features
or hardware required for drawing them is not present or not
enabled.