Moved mAh/mi and added & to fix OSD menu
It appears that the OSD menu has always had a bug. From the old fonts, I can see that the ampersand symbol was never present. It is used in the OSD menu. This is now fixed. I have created a temporary & symbol on index 038. This will be replaced with the correct symbol for each font.