Changed GPS Speed to Ground Speed
This change is to make the term for velocity along the ground consistent with aeronautical terms. It also makes it consistent with the DJI speed selector drop down on the right side of the page.

I have also slightly modified the DJI labels. Using italics to highlight the sensor name in the DJI OSD setup.