Update failsafe.html
Reverted to the original code as the failsafe page uses a different data method at the moment. This means no conversion, but it will works as it was for the time being. This can be updated soon to have the conversion. Time is the issue at the moment. While here, I took the liberty of correcting the max for both failsafe_delay and failsafe_off_delay.