Adjusted 4:3 preview
To me, this looks wrong. So I have left the calculated values in there, commented out.

I tested in the Avatar goggles with the camera set to 4:3. Then moved the guides to match. Unfortunately I don't have the correct adapter to take the USB out of the Avatar goggles and in to my FatShark HDO goggles (4:3). I will get hold of an adapter and re test. But this can be updated through a bugfix PR if necessary. I've also read that even if set to 4:3, the VRx/goggles still output 16:9. So it may even need a firmware update on the Walksnail side to get this 100% confirmed.