Fix profile change not working
Initially, I thought that the profiles were not changing when you selected a new profile from the select box. It turns out that this was only happening after setting up from the defaults dialog. The problem was that the `periodicStatusUpdater` was being stopped to perform the parameter upload. But not re-started. I added a function to resume the updater, and everything works as it should. I made a few other changes:
- Tidied up the setting of the parameters for each profile
- Added specific `set` profile messages, rather than `loaded` being repeated
- Only show messages for the profiles that have changed, not all of them